REFACTOR : MemberController return resource 경로 리팩토링
